# swagger.json
{
...
  "info": {
    "title": "API Documentation",
    "version": "0.0.1"
  },
  "paths": {
    "/admin/balance": {
      "put": {
        "description": "<br/>The endpoint will update the user balance for the current Calendar Period only. If the user balance needs to updated permanently then it is recommended to update the user's balance via the admin PUT user endpoint.<br/>",
        "parameters": [
          {
            "description": "Admin API Key for authentication",
            "in": "header",
            "name": "API-Key",
            "required": true,
            "type": "string"
          },
          {
            "description": "Valid token for verification",
            "in": "query",
            "name": "token",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "new_balance": {
                  "description": "New balance value",
                  "type": "integer"
                },
                "user_id": {
                  "description": "ID of user to update",
                  "type": "string"
                }
              },
              "required": [
                "user_id",
                "new_balance"
              ],
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Balance updated successfully"
          },
          "401": {
            "description": "Authentication error"
          },
          "403": {
            "description": "Forbidden - not an admin"
          },
          "500": {
            "description": "Server error"
          }
        },
        "summary": "Update user's API call balance for the current Calendar Period (Admin only). ",
        "tags": [
          "Admin Functions"
        ]
      }
    },
...
    "/token": {
      "get": {
        "parameters": [
          {
            "description": "API Key for authentication",
            "in": "header",
            "name": "API-Key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Token generated successfully",
            "schema": {
              "properties": {
                "access_token": {
                  "description": "generated access token to use with api calls",
                  "type": "string"
                },
                "expires_in": {
                  "description": "Time in seconds until token expiration",
                  "format": "seconds",
                  "type": "integer"
                },
                "expires_on": {
                  "description": "Token expiration timestamp",
                  "format": "date-time",
                  "type": "string"
                },
                "token_type": {
                  "description": "Type of token generated",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "401": {
            "description": "Authentication error",
            "schema": {
              "properties": {
                "error": {
                  "example": "Authentication Error",
                  "type": "string"
                },
                "message": {
                  "example": "Missing API Key header (API-Key) or Invalid API Key",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "properties": {
                "error": {
                  "example": "Server Error",
                  "type": "string"
                },
                "message": {
                  "example": "Error generating token",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Generate a token for API access using a valid api authentication key.",
        "tags": [
          "Token Service"
        ]
      }
    },
...
    "/usage/balance": {
      "get": {
        "parameters": [
          {
            "description": "API Key for authentication",
            "in": "header",
            "name": "API-Key",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Current balance retrieved successfully"
          },
          "401": {
            "description": "Authentication error"
          },
          "500": {
            "description": "Server error"
          }
        },
        "summary": "Check current API call balance for authenticated user",
        "tags": [
          "Balance Management"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "description": "API Key for authentication. Add your token with the X-Token header.",
      "in": "header",
      "name": "X-Token",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}